Crate polars_io

Re-exports

pub use cloud::glob as async_glob;cloud

pub use path_utils::*;

Modules

avroavro

catalogcatalog

cloud

Interface with cloud storage through the object_store crate.

configs

csvcsv or json

Functionality for reading and writing CSV files.

file_cachefile_cache

hive

ipcipc_streaming or ipc

jsonjson

(De)serialize JSON files.

metrics

mmap

ndjsonjson

parquetparquet

Functionality for reading and writing Apache Parquet files.

path_utils

pl_asyncasync

predicates

prelude

scan_linesscan_lines

utils

Macros

impl_page_walkcatalog

Support for traversing paginated response values that look like:

Structs

HiveOptions

Options for Hive partitioning.

RowIndex

Enums

ExternalCompression

Compression options for file that are expressed externally like CSV and NDJSON. Externally does not mean by an external tool, more that it doesn’t happen internally like it does for Parquet and IPC.

Traits

ArrowReader

SerReader

SerWriter

Functions

schema_to_arrow_checked