Enum polars_io::cloud::options::AzureConfigKey

#[non_exhaustive]
pub enum AzureConfigKey {

    AccountName,
    AccessKey,
    ClientId,
    ClientSecret,
    AuthorityId,
    SasKey,
    Token,
    UseEmulator,
    Endpoint,
    UseFabricEndpoint,
    MsiEndpoint,
    ObjectId,
    MsiResourceId,
    FederatedTokenFile,
    UseAzureCli,
    SkipSignature,
    ContainerName,
    DisableTagging,
    Client(ClientConfigKey),

}

Available on crate feature azure only.

Configuration keys for [MicrosoftAzureBuilder]

Configuration via keys can be done via [MicrosoftAzureBuilder::with_config]

Example

let builder = MicrosoftAzureBuilder::new()
    .with_config("azure_client_id".parse().unwrap(), "my-client-id")
    .with_config(AzureConfigKey::AuthorityId, "my-tenant-id");

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccountName

The name of the azure storage account

Supported keys:

azure_storage_account_name

account_name

AccessKey

Master key for accessing storage account

Supported keys:

azure_storage_account_key

azure_storage_access_key

azure_storage_master_key

access_key

account_key

master_key

ClientId

Service principal client id for authorizing requests

Supported keys:

azure_storage_client_id

azure_client_id

client_id

ClientSecret

Service principal client secret for authorizing requests

Supported keys:

azure_storage_client_secret

azure_client_secret

client_secret

AuthorityId

Tenant id used in oauth flows

Supported keys:

azure_storage_tenant_id

azure_storage_authority_id

azure_tenant_id

azure_authority_id

tenant_id

authority_id

SasKey

Shared access signature.

The signature is expected to be percent-encoded, much like they are provided in the azure storage explorer or azure portal.

Supported keys:

azure_storage_sas_key

azure_storage_sas_token

sas_key

sas_token

Token

Bearer token

Supported keys:

azure_storage_token

bearer_token

token

UseEmulator

Use object store with azurite storage emulator

Supported keys:

azure_storage_use_emulator

object_store_use_emulator

use_emulator

Endpoint

Override the endpoint used to communicate with blob storage

Supported keys:

azure_storage_endpoint

azure_endpoint

endpoint

UseFabricEndpoint

Use object store with url scheme account.dfs.fabric.microsoft.com

Supported keys:

azure_use_fabric_endpoint

use_fabric_endpoint

MsiEndpoint

Endpoint to request a imds managed identity token

Supported keys:

azure_msi_endpoint

azure_identity_endpoint

identity_endpoint

msi_endpoint

ObjectId

Object id for use with managed identity authentication

Supported keys:

azure_object_id

object_id

MsiResourceId

Msi resource id for use with managed identity authentication

Supported keys:

azure_msi_resource_id

msi_resource_id

FederatedTokenFile

File containing token for Azure AD workload identity federation

Supported keys:

azure_federated_token_file

federated_token_file

UseAzureCli

Use azure cli for acquiring access token

Supported keys:

azure_use_azure_cli

use_azure_cli

SkipSignature

Skip signing requests

Supported keys:

azure_skip_signature

skip_signature

ContainerName

Container name

Supported keys:

azure_container_name

container_name

DisableTagging

Disables tagging objects

This can be desirable if not supported by the backing store

Supported keys:

azure_disable_tagging

disable_tagging

Client(ClientConfigKey)

Client options