polars_io::csv::read

Struct CsvReadOptions

pub struct CsvReadOptions {

    pub path: Option<PathBuf>,
    pub rechunk: bool,
    pub n_threads: Option<usize>,
    pub low_memory: bool,
    pub n_rows: Option<usize>,
    pub row_index: Option<RowIndex>,
    pub columns: Option<Arc<[PlSmallStr]>>,
    pub projection: Option<Arc<Vec<usize>>>,
    pub schema: Option<SchemaRef>,
    pub schema_overwrite: Option<SchemaRef>,
    pub dtype_overwrite: Option<Arc<Vec<DataType>>>,
    pub parse_options: Arc<CsvParseOptions>,
    pub has_header: bool,
    pub chunk_size: usize,
    pub skip_rows: usize,
    pub skip_lines: usize,
    pub skip_rows_after_header: usize,
    pub infer_schema_length: Option<usize>,
    pub raise_if_empty: bool,
    pub ignore_errors: bool,
    pub fields_to_cast: Vec<Field>,

}

Available on crate features csv or json only.

Fields
path: Option<PathBuf>rechunk: booln_threads: Option<usize>low_memory: booln_rows: Option<usize>row_index: Option<RowIndex>columns: Option<Arc<[PlSmallStr]>>projection: Option<Arc<Vec<usize>>>schema: Option<SchemaRef>schema_overwrite: Option<SchemaRef>dtype_overwrite: Option<Arc<Vec<DataType>>>parse_options: Arc<CsvParseOptions>has_header: boolchunk_size: usizeskip_rows: usize

Skip rows according to the CSV spec.

skip_lines: usize

Skip lines according to newline char (e.g. escaping will be ignored)

skip_rows_after_header: usizeinfer_schema_length: Option<usize>raise_if_empty: boolignore_errors: boolfields_to_cast: Vec<Field>
Implementations

impl CsvReadOptions

pub fn get_parse_options(&self) -> Arc<CsvParseOptions>

pub fn with_path<P: Into<PathBuf>>(self, path: Option<P>) -> Self

pub fn with_rechunk(self, rechunk: bool) -> Self

Whether to makes the columns contiguous in memory.

pub fn with_n_threads(self, n_threads: Option<usize>) -> Self

Number of threads to use for reading. Defaults to the size of the polars thread pool.

pub fn with_low_memory(self, low_memory: bool) -> Self

Reduce memory consumption at the expense of performance

pub fn with_n_rows(self, n_rows: Option<usize>) -> Self

Limits the number of rows to read.

pub fn with_row_index(self, row_index: Option<RowIndex>) -> Self

Adds a row index column.

pub fn with_columns(self, columns: Option<Arc<[PlSmallStr]>>) -> Self

Which columns to select.

pub fn with_projection(self, projection: Option<Arc<Vec<usize>>>) -> Self

Which columns to select denoted by their index. The index starts from 0 (i.e. [0, 4] would select the 1st and 5th column).

pub fn with_schema(self, schema: Option<SchemaRef>) -> Self

Set the schema to use for CSV file. The length of the schema must match the number of columns in the file. If this is None, the schema is inferred from the file.

pub fn with_schema_overwrite(self, schema_overwrite: Option<SchemaRef>) -> Self

Overwrites the data types in the schema by column name.

pub fn with_dtype_overwrite( self, dtype_overwrite: Option<Arc<Vec<DataType>>>, ) -> Self

Overwrite the dtypes in the schema in the order of the slice that’s given. This is useful if you don’t know the column names beforehand

pub fn with_parse_options(self, parse_options: CsvParseOptions) -> Self

Sets the CSV parsing options. See map_parse_options for an easier way to mutate them in-place.

pub fn with_has_header(self, has_header: bool) -> Self

Sets whether the CSV file has a header row.

pub fn with_chunk_size(self, chunk_size: usize) -> Self

Sets the chunk size used by the parser. This influences performance.

pub fn with_skip_rows(self, skip_rows: usize) -> Self

Start reading after skip_rows rows. The header will be parsed at this offset. Note that we respect CSV escaping/comments when skipping rows. If you want to skip by newline char only, use skip_lines.

pub fn with_skip_lines(self, skip_lines: usize) -> Self

Start reading after skip_lines lines. The header will be parsed at this offset. Note that CSV escaping will not be respected when skipping lines. If you want to skip valid CSV rows, use skip_rows.

pub fn with_skip_rows_after_header(self, skip_rows_after_header: usize) -> Self

Number of rows to skip after the header row.

pub fn with_infer_schema_length( self, infer_schema_length: Option<usize>, ) -> Self

Set the number of rows to use when inferring the csv schema. The default is 100 rows. Setting to None will do a full table scan, which is very slow.

pub fn with_raise_if_empty(self, raise_if_empty: bool) -> Self

Whether to raise an error if the frame is empty. By default an empty DataFrame is returned.

pub fn with_ignore_errors(self, ignore_errors: bool) -> Self

Continue with next batch when a ParserError is encountered.

pub fn map_parse_options<F: Fn(CsvParseOptions) -> CsvParseOptions>( self, map_func: F, ) -> Self

Apply a function to the parse options.

impl CsvReadOptions

pub fn try_into_reader_with_file_path( self, path: Option<PathBuf>, ) -> PolarsResult<CsvReader<File>>

Creates a CSV reader using a file path.

Panics

If both self.path and the path parameter are non-null. Only one of them is to be non-null.

pub fn into_reader_with_file_handle<R: MmapBytesReader>( self, reader: R, ) -> CsvReader<R>

Creates a CSV reader using a file handle.

impl CsvReadOptions

pub fn update_with_inference_result( &mut self, si_result: &SchemaInferenceResult, )

Note: This does not update the schema from the inference result.