Struct polars_io::csv::read::CsvReader

pub struct CsvReader<'a, R>
where
    R: MmapBytesReader,
{ /* private fields */ }

Available on crate features csv or json only.

Create a new DataFrame by reading a csv file.

Example

use polars_core::prelude::*;
use polars_io::prelude::*;
use std::fs::File;

fn example() -> PolarsResult<DataFrame> {
    CsvReader::from_path("iris.csv")?
            .has_header(true)
            .finish()
}

Implementations

impl<'a, R> CsvReader<'a, R>
where R: 'a + MmapBytesReader,

pub fn with_options(self, options: CsvReaderOptions) -> Self

Skip these rows after the header

pub fn has_header(self, has_header: bool) -> Self

Sets whether the CSV file has headers

pub fn with_separator(self, separator: u8) -> Self

Sets the CSV file’s column separator as a byte character

pub fn with_quote_char(self, quote_char: Option<u8>) -> Self

Sets the char used as quote char. The default is b'"'. If set to None, quoting is disabled.

pub fn with_comment_prefix(self, comment_prefix: Option<&str>) -> Self

Sets the comment prefix for this instance. Lines starting with this prefix will be ignored.

pub fn _with_comment_prefix(self, comment_prefix: Option<CommentPrefix>) -> Self

Sets the comment prefix from CsvParserOptions for internal initialization.

pub fn with_end_of_line_char(self, eol_char: u8) -> Self

Set the char used as end-of-line char. The default is b'\n'.

pub fn with_encoding(self, encoding: CsvEncoding) -> Self

Set CsvEncoding.

pub fn with_skip_rows(self, n: usize) -> Self

Skip the first n rows during parsing. The header will be parsed at n lines.

pub fn with_skip_rows_after_header(self, n: usize) -> Self

Skip these rows after the header

pub fn with_schema(self, schema: Option<SchemaRef>) -> Self

Set the CSV file’s schema. This only accepts datatypes that are implemented in the csv parser and expects a complete Schema.

It is recommended to use with_dtypes instead.

pub fn with_dtypes(self, schema: Option<SchemaRef>) -> Self

Overwrite the schema with the dtypes in this given Schema. The given schema may be a subset of the total schema.

pub fn infer_schema(self, n: Option<usize>) -> Self

Set the CSV reader to infer the schema of the file

Arguments

n - Maximum number of rows read for schema inference. Setting this to None will do a full table scan (slow).

pub fn with_try_parse_dates(self, toggle: bool) -> Self

Automatically try to parse dates/ datetimes and time. If parsing fails, columns remain of dtype [DataType::String].

pub fn with_null_values(self, null_values: Option<NullValues>) -> Self

Set values that will be interpreted as missing/null.

Note: any value you set as null value will not be escaped, so if quotation marks are part of the null value you should include them.

pub fn with_ignore_errors(self, toggle: bool) -> Self

Continue with next batch when a ParserError is encountered.

pub fn raise_if_empty(self, toggle: bool) -> Self

Raise an error if CSV is empty (otherwise return an empty frame)

pub fn truncate_ragged_lines(self, toggle: bool) -> Self

Truncate lines that are longer than the schema.

pub fn with_decimal_comma(self, toggle: bool) -> Self

Parse floats with a comma as decimal separator.

pub fn with_n_threads(self, n: Option<usize>) -> Self

Set the number of threads used in CSV reading. The default uses the number of cores of your cpu.

Note that this only works if this is initialized with CsvReader::from_path. Note that the number of cores is the maximum allowed number of threads.

pub fn low_memory(self, toggle: bool) -> Self

Reduce memory consumption at the expense of performance

pub fn with_row_index(self, row_index: Option<RowIndex>) -> Self

Add a row index column.

pub fn with_chunk_size(self, chunk_size: usize) -> Self

Sets the chunk size used by the parser. This influences performance

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_rechunk(self, rechunk: bool) -> Self

Rechunk the DataFrame to contiguous memory after the CSV is parsed.

pub fn with_missing_is_null(self, missing_is_null: bool) -> Self

Treat missing fields as null.

pub fn with_dtypes_slice(self, dtypes: Option<&'a [DataType]>) -> Self

Overwrite the dtypes in the schema in the order of the slice that’s given. This is useful if you don’t know the column names beforehand

pub fn with_projection(self, projection: Option<Vec<usize>>) -> Self

Set the reader’s column projection. This counts from 0, meaning that vec![0, 4] would select the 1st and 5th column.

pub fn with_columns(self, columns: Option<Vec<String>>) -> Self

Columns to select/ project

pub fn with_path<P: Into<PathBuf>>(self, path: Option<P>) -> Self

The preferred way to initialize this builder. This allows the CSV file to be memory mapped and thereby greatly increases parsing performance.

pub fn sample_size(self, size: usize) -> Self

Sets the size of the sample taken from the CSV file. The sample is used to get statistic about the file. These statistics are used to try to optimally allocate up front. Increasing this may improve performance.

pub fn with_predicate(self, predicate: Option<Arc<dyn PhysicalIoExpr>>) -> Self

impl<'a> CsvReader<'a, File>

pub fn from_path<P: Into<PathBuf>>(path: P) -> PolarsResult<Self>

This is the recommended way to create a csv reader as this allows for fastest parsing.

impl<'a, R: MmapBytesReader + 'a> CsvReader<'a, R>

pub fn batched_borrowed_mmap( &'a mut self ) -> PolarsResult<BatchedCsvReaderMmap<'a>>

pub fn batched_borrowed_read( &'a mut self ) -> PolarsResult<BatchedCsvReaderRead<'a>>

impl<'a> CsvReader<'a, Box<dyn MmapBytesReader>>

pub fn batched_mmap( self, schema: Option<SchemaRef> ) -> PolarsResult<OwnedBatchedCsvReaderMmap>

pub fn batched_read( self, schema: Option<SchemaRef> ) -> PolarsResult<OwnedBatchedCsvReader>