polars_io

Module hive

Functions

merge_sorted_to_schema_order

Merge 2 lists of columns into one, where each list contains columns ordered such that their indices in the schema are in ascending order.

merge_sorted_to_schema_order_impl