Struct polars_io::ndjson::core::JsonLineReader

pub struct JsonLineReader<'a, R>
where
    R: MmapBytesReader,
{ /* private fields */ }

Available on crate feature json only.

Implementations

impl<'a, R> JsonLineReader<'a, R>
where R: 'a + MmapBytesReader,

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

pub fn with_schema(self, schema: SchemaRef) -> Self

pub fn with_schema_overwrite(self, schema: &'a Schema) -> Self

pub fn with_rechunk(self, rechunk: bool) -> Self

pub fn infer_schema_len(self, infer_schema_len: Option<NonZeroUsize>) -> Self

pub fn with_n_threads(self, n: Option<usize>) -> Self

pub fn with_path<P: Into<PathBuf>>(self, path: Option<P>) -> Self

pub fn with_chunk_size(self, chunk_size: Option<NonZeroUsize>) -> Self

Sets the chunk size used by the parser. This influences performance

pub fn low_memory(self, toggle: bool) -> Self

Reduce memory consumption at the expense of performance

pub fn with_ignore_errors(self, ignore_errors: bool) -> Self

Set values as Null if parsing fails because of schema mismatches.

impl<'a> JsonLineReader<'a, File>

pub fn from_path<P: Into<PathBuf>>(path: P) -> PolarsResult<Self>

This is the recommended way to create a json reader as this allows for fastest parsing.