polars_io::ndjson::core

Struct StructArray

pub struct StructArray { /* private fields */ }

Available on crate feature json only.

A StructArray is a nested [Array] with an optional validity representing multiple [Array] with the same number of rows.

Example

use polars_arrow::array::*;
use polars_arrow::datatypes::*;
let boolean = BooleanArray::from_slice(&[false, false, true, true]).boxed();
let int = Int32Array::from_slice(&[42, 28, 19, 31]).boxed();

let fields = vec![
    Field::new("b".into(), ArrowDataType::Boolean, false),
    Field::new("c".into(), ArrowDataType::Int32, false),
];

let array = StructArray::new(ArrowDataType::Struct(fields), 4, vec![boolean, int], None);

Implementations

impl<'a> StructArray

pub fn iter( &'a self, ) -> ZipValidity<Vec<Box<dyn Scalar>>, StructValueIter<'a>, BitmapIter<'a>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> StructValueIter<'a>

Returns an iterator of Box<dyn Array>

impl StructArray

pub fn try_new( dtype: ArrowDataType, length: usize, values: Vec<Box<dyn Array>>, validity: Option<Bitmap>, ) -> Result<StructArray, PolarsError>

Returns a new StructArray.

Errors

This function errors iff:

dtype’s physical type is not [crate::datatypes::PhysicalType::Struct].

the children of dtype are empty

the values’s len is different from children’s length

any of the values’s data type is different from its corresponding children’ data type

any element of values has a different length than the first element

the validity’s length is not equal to the length of the first element

pub fn new( dtype: ArrowDataType, length: usize, values: Vec<Box<dyn Array>>, validity: Option<Bitmap>, ) -> StructArray

Returns a new StructArray

Panics

This function panics iff:

dtype’s physical type is not [crate::datatypes::PhysicalType::Struct].

the children of dtype are empty

the values’s len is different from children’s length

any of the values’s data type is different from its corresponding children’ data type

any element of values has a different length than the first element

the validity’s length is not equal to the length of the first element

pub fn new_empty(dtype: ArrowDataType) -> StructArray

Creates an empty StructArray.

pub fn new_null(dtype: ArrowDataType, length: usize) -> StructArray

Creates a null StructArray of length length.

impl StructArray

pub fn into_data( self, ) -> (Vec<Field>, usize, Vec<Box<dyn Array>>, Option<Bitmap>)

Deconstructs the StructArray into its individual components.

pub fn slice(&mut self, offset: usize, length: usize)

Slices this StructArray.

Panics

panics iff offset + length > self.len()

Implementation

This operation is O(F) where F is the number of fields.

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this StructArray.

Implementation

This operation is O(F) where F is the number of fields.

Safety

The caller must ensure that offset + length <= self.len().

pub fn sliced(self, offset: usize, length: usize) -> StructArray

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn sliced_unchecked( self, offset: usize, length: usize, ) -> StructArray

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> StructArray

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn take_validity(&mut self) -> Option<Bitmap>

Takes the validity of this array, leaving it without a validity mask.

pub fn boxed(self) -> Box<dyn Array>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

impl StructArray

pub fn len(&self) -> usize

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn values(&self) -> &[Box<dyn Array>]

Returns the values of this StructArray.

pub fn fields(&self) -> &[Field]

Returns the fields of this StructArray.

impl StructArray

pub fn get_fields(dtype: &ArrowDataType) -> &[Field]

Returns the fields the DataType::Struct.