polars_io::parquet
Module read

Available on crate feature parquet only.

Functionality for reading Apache Parquet files.

Examples

use polars_core::prelude::*;
use polars_io::prelude::*;
use std::fs::File;

fn example() -> PolarsResult<DataFrame> {
    let r = File::open("example.parquet").unwrap();
    let reader = ParquetReader::new(r);
    reader.finish()
}

Modules

_internal

Structs

BatchedParquetReader

FileMetadata

Metadata for a Parquet file.

ParquetAsyncReadercloud

A Parquet reader on top of the async object_store API. Only the batch reader is implemented since parquet files on cloud storage tend to be big and slow to access.

ParquetOptions

ParquetReader

Read Apache parquet format into a DataFrame.

Enums

ParallelStrategy

Functions

create_sorting_map

infer_schema

Infers a [ArrowSchema] from parquet’s FileMetadata.

materialize_empty_df

try_set_sorted_flag