polars_io::parquet::read
Module _internal

Available on crate feature parquet only.

Enums

PrefilterMaskSetting

Functions

calc_prefilter_cost

ensure_matching_dtypes_if_found

Utility to ensure the dtype of the column in current_schema matches the dtype in schema if that column exists in schema.

read_this_row_group

to_deserializer