polars_io::parquet

Module write

Available on crate feature parquet only.

Functionality for reading and writing Apache Parquet files.

Structs

BatchedWriter

BrotliLevel

A valid Brotli compression level.

GzipLevel

A valid Gzip compression level.

MetadataKeyValue

ParquetFieldOverwrites

ParquetMetadataContext

Context that can be used to construct custom file-level key value metadata for a Parquet file.

ParquetWriteOptions

ParquetWriter

Write a DataFrame to Parquet format.

StatisticsOptions

The statistics to write

ZstdLevel

A valid Zstandard compression level.

Enums

ChildFieldOverwrites

KeyValueMetadata

Key/value pairs that can be attached to a Parquet file as file-level metadtaa.

ParquetCompression

The compression strategy to use for writing Parquet files.

Functions

get_column_write_options

Type Aliases

RowGroupIterColumns