polars_io::path_utils

Function expand_paths_hive

pub async fn expand_paths_hive(
    paths: &[PlRefPath],
    glob: bool,
    hidden_file_prefix: &[PlSmallStr],
    cloud_options: &mut Option<CloudOptions>,
    check_directory_level: bool,
) -> PolarsResult<(Buffer<PlRefPath>, usize)>

Recursively traverses directories and expands globs if glob is true. Returns the expanded paths and the index at which to start parsing hive partitions from the path.