Struct polars_io::predicates::ColumnStats

pub struct ColumnStats { /* private fields */ }

Statistics of the values in a column.

The following statistics are tracked for each row group:

Null count

Minimum value

Maximum value

Implementations

impl ColumnStats

pub fn new( field: Field, null_count: Option<Series>, min_value: Option<Series>, max_value: Option<Series> ) -> Self

Constructs a new ColumnStats.

pub fn from_field(field: Field) -> Self

Constructs a new ColumnStats with only the Field information and no statistics.

pub fn from_column_literal(s: Series) -> Self

Constructs a new ColumnStats from a single-value Series.

pub fn dtype(&self) -> &DataType

Returns the DataType of the column.

pub fn get_null_count_state(&self) -> Option<&Series>

Returns the null count of each row group of the column.

pub fn get_min_state(&self) -> Option<&Series>

Returns the minimum value of each row group of the column.

pub fn get_max_state(&self) -> Option<&Series>

Returns the maximum value of each row group of the column.

pub fn null_count(&self) -> Option<usize>

Returns the null count of the column.

pub fn to_min_max(&self) -> Option<Series>

Returns the minimum and maximum values of the column as a single Series.

pub fn to_min(&self) -> Option<&Series>

Returns the minimum value of the column as a single-value Series.

Returns None if no maximum value is available.

pub fn to_max(&self) -> Option<&Series>

Returns the maximum value of the column as a single-value Series.

Returns None if no maximum value is available.