Trait polars_io::SerReader

pub trait SerReader<R>where
    R: Read,{
    // Required methods
    fn new(reader: R) -> Self;
    fn finish(self) -> PolarsResult<DataFrame>;

    // Provided method
    fn set_rechunk(self, _rechunk: bool) -> Self
       where Self: Sized { ... }
}

Required Methods

fn new(reader: R) -> Self

Create a new instance of the [SerReader]

fn finish(self) -> PolarsResult<DataFrame>

Take the SerReader and return a parsed DataFrame.

Provided Methods

fn set_rechunk(self, _rechunk: bool) -> Selfwhere Self: Sized,

Make sure that all columns are contiguous in memory by aggregating the chunks into a single array.