polars_io
Module utils

Modules

byte_sourcecloud

compression

file

slice

Constants

URL_ENCODE_CHAR_SET

Statics

BOOLEAN_RE

FLOAT_RE

FLOAT_RE_DECIMAL

INTEGER_RE

Functions

apply_projectionipc or ipc_streaming or parquet or avro

columns_to_projectionipc or ipc_streaming or avro or parquet

decode_json_responsecloud

Utility for decoding JSON that adds the response value to the error message if decoding fails. This makes it much easier to debug errors from parsing network responses.

get_reader_bytes

materialize_projection

overwrite_schemajson