polars_io::utils

Module compression

Enums

ByteSourceReader

A byte source that abstracts over in-memory buffers and streaming readers, with optional transparent decompression and buffering.

CompressedReader

Reader that implements a streaming read trait for uncompressed, gzip, zlib and zstd compression.

CompressedWriter

Constructor for WriteableTrait compressed encoders.

SupportedCompression

Represents the compression algorithms that we have decoders for

Functions

maybe_decompress_bytesDeprecated

Decompress bytes if compression is detected, otherwise simply return it. An out vec must be given for ownership of the decompressed data.