polars_io::utils::compression

Enum CompressedReader

pub enum CompressedReader {
    Uncompressed {
        slice: Buffer<u8>,
        offset: usize,
    },
    Gzip(MultiGzDecoder<Cursor<Buffer<u8>>>),
    Zlib(ZlibDecoder<Cursor<Buffer<u8>>>),
    Zstd(Decoder<'static, Cursor<Buffer<u8>>>),
}

Reader that implements a streaming read trait for uncompressed, gzip, zlib and zstd compression.

This allows handling decompression transparently in a streaming fashion.

Variants

Uncompressed

Fields

slice: Buffer<u8>

offset: usize

Gzip(MultiGzDecoder<Cursor<Buffer<u8>>>)

Available on crate feature decompress only.

Zlib(ZlibDecoder<Cursor<Buffer<u8>>>)

Available on crate feature decompress only.

Zstd(Decoder<'static, Cursor<Buffer<u8>>>)

Available on crate feature decompress only.

Implementations

impl CompressedReader

pub fn try_new(slice: Buffer<u8>) -> PolarsResult<Self>

pub fn is_compressed(&self) -> bool

pub const fn initial_read_size() -> usize

pub const fn ideal_read_size() -> usize

pub fn total_len_estimate(&self) -> usize

If possible returns the total number of bytes that will be produced by reading from the start to finish.

pub fn read_next_slice( &mut self, prev_leftover: &Buffer<u8>, read_size: usize, ) -> Result<(Buffer<u8>, usize)>

Reads exactly read_size bytes if possible from the internal readers and creates a new [Buffer] with the content concat(prev_leftover, new_bytes).

Returns the new slice and the number of bytes read, which will be 0 when eof is reached and this function is called again.

If the underlying reader is uncompressed the operation is a cheap zero-copy [Buffer::sliced] operation.

By handling slice concatenation at this level we can implement zero-copy reading and make the interface easier to use.

Itโs a logic bug if prev_leftover is neither empty nor the last slice returned by this function.