polars_io::utils::compression

Function maybe_decompress_bytes

pub fn maybe_decompress_bytes<'a>(
    bytes: &'a [u8],
    out: &'a mut Vec<u8>,
) -> PolarsResult<&'a [u8]>

Decompress bytes if compression is detected, otherwise simply return it. An out vec must be given for ownership of the decompressed data.