polars_io
Module cloud

Interface with cloud storage through the object_store crate.

Re-exports

pub use options::*;

Modules

credential_providercloud

options

Structs

CloudLocationcloud

A location on cloud storage, may have wildcards.

CloudWritercloud

Adaptor which wraps the interface of ObjectStore::BufWriter exposing a synchronous interface which implements std::io::Write.

PolarsObjectStorecloud

Polars specific wrapper for Arc<dyn ObjectStore> that limits the number of concurrent requests for the entire application.

Functions

build_object_storecloud

Build an [ObjectStore] based on the URL and passed in url. Return the cloud location and an implementation of the object store.

globcloud

List files with a prefix derived from the pattern.

object_path_from_strcloud

Construct an object_store Path from a string without any encoding/decoding.

Type Aliases

ObjectStorePathcloud