polars_io

Module cloud

Interface with cloud storage through the object_store crate.

Re-exports

pub use options::*;

Modules

credential_providercloud

options

Structs

BlockingCloudWritercloud

Adaptor which wraps the interface of ObjectStore::BufWriter exposing a synchronous interface which implements std::io::Write.

CloudLocationcloud

A location on cloud storage, may have wildcards.

PolarsObjectStorecloud

Polars wrapper around [ObjectStore] functionality. This struct is cheaply cloneable.

Functions

build_object_storecloud

Build an [ObjectStore] based on the URL and passed in url. Return the cloud location and an implementation of the object store.

globcloud

List files with a prefix derived from the pattern.

object_path_from_strcloud

Construct an object_store Path from a string without any encoding/decoding.

Type Aliases

ObjectStorePathcloud