polars_io::csv

Module read

Available on crate features csv or json only.

Functionality for reading CSV files.

Examples

use polars_core::prelude::*;
use polars_io::prelude::*;
use std::fs::File;

fn example() -> PolarsResult<DataFrame> {
    // Prefer `from_path` over `new` as it is faster.
    CsvReadOptions::default()
        .with_has_header(true)
        .try_into_reader_with_file_path(Some("example.csv".into()))?
        .finish()
}

Re-exports

pub use schema_inference::infer_file_schema;

Modules

_csv_read_internal

buffer

schema_inference

Structs

BatchedCsvReader

CsvParseOptions

CsvReadOptions

CsvReader

Create a new DataFrame by reading a csv file.

OwnedBatchedCsvReader

Enums

CommentPrefix

CsvEncoding

NullValues

Functions

count_rows

Read the number of rows without parsing columns useful for count(*) queries

count_rows_from_slice

Read the number of rows without parsing columns

count_rows_from_slice_par

Read the number of rows without parsing columns useful for count(*) queries