Function polars_io::csv::utils::infer_file_schema

pub fn infer_file_schema(
    reader_bytes: &ReaderBytes<'_>,
    delimiter: u8,
    max_read_lines: Option<usize>,
    has_header: bool,
    schema_overwrite: Option<&Schema>,
    skip_rows: &mut usize,
    skip_rows_after_header: usize,
    comment_char: Option<u8>,
    quote_char: Option<u8>,
    eol_char: u8,
    null_values: Option<&NullValues>,
    try_parse_dates: bool
) -> PolarsResult<(Schema, usize, usize)>

Available on (crate features csv-file or json) and crate feature private only.

Infer the schema of a CSV file by reading through the first n records of the file, with max_read_records controlling the maximum number of records to read.

If max_read_records is not set, the whole file is read to infer its schema.

Returns - inferred schema - number of rows used for inference. - bytes read