polars_io::hive

Function merge_sorted_to_schema_order

pub fn merge_sorted_to_schema_order<'a, D>(
    cols_lhs: &'a mut dyn Iterator<Item = Column>,
    cols_rhs: &'a mut dyn Iterator<Item = Column>,
    schema: &Schema<D>,
    output: &'a mut Vec<Column>,
)

Merge 2 lists of columns into one, where each list contains columns ordered such that their indices in the schema are in ascending order.

Layouts:

cols_lhs: [row_index?, ..schema_columns?, ..other_left?]

If the first item in cols_lhs is not found in the schema, it will be assumed to be a row_index column and placed first into the result.

cols_rhs: [..schema_columns? ..other_right?]

Output:

[..schema_columns?, ..other_left?, ..other_right?]

Note: The row_index column should be handled before calling this function.

Panics

Panics if either cols_lhs or cols_rhs is empty.