polars_io::mmap

Trait MmapBytesReader

pub trait MmapBytesReader:
    Read
    + Seek
    + Send
    + Sync {
    // Provided methods
    fn to_file(&self) -> Option<&File> { ... }
    fn to_bytes(&self) -> Option<&[u8]> { ... }
}

Trait used to get a hold to file handler or to the underlying bytes without performing a Read.

Provided Methods

fn to_file(&self) -> Option<&File>

fn to_bytes(&self) -> Option<&[u8]>