Struct polars_io::parquet::ParquetReader

pub struct ParquetReader<R: Read + Seek> { /* private fields */ }

Read Apache parquet format into a DataFrame.

Implementations

impl<R: MmapBytesReader> ParquetReader<R>

pub fn _finish_with_scan_ops(
self,
predicate: Option<Arc<dyn PhysicalIoExpr>>,
projection: Option<&[usize]>
) -> PolarsResult<DataFrame>

pub fn set_low_memory(self, low_memory: bool) -> Self

Try to reduce memory pressure at the expense of performance. If setting this does not reduce memory enough, turn off parallelization.

pub fn read_parallel(self, parallel: ParallelStrategy) -> Self

Read the parquet file in parallel (default). The single threaded reader consumes less memory.

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

Stop parsing when n rows are parsed. By settings this parameter the csv will be parsed sequentially.

pub fn with_columns(self, columns: Option<Vec<String>>) -> Self

Columns to select/ project

pub fn with_projection(self, projection: Option<Vec<usize>>) -> Self

Set the reader’s column projection. This counts from 0, meaning that vec![0, 4] would select the 1st and 5th column.

pub fn with_row_count(self, row_count: Option<RowCount>) -> Self

Add a row_count column.

pub fn schema(&mut self) -> PolarsResult<Schema>

Schema of the file.

pub fn num_rows(&mut self) -> PolarsResult<usize>

Number of rows in the parquet file.

impl<R: MmapBytesReader + 'static> ParquetReader<R>

pub fn batched(self, chunk_size: usize) -> PolarsResult<BatchedParquetReader>