polars_io
Module path_utils

Statics

POLARS_TEMP_DIR_BASE_PATH

Functions

expand_paths

Recursively traverses directories and expands globs if glob is true.

expand_paths_hive

Recursively traverses directories and expands globs if glob is true. Returns the expanded paths and the index at which to start parsing hive partitions from the path.

expanded_from_single_directory

Returns true if expanded_paths were expanded from a single directory

get_glob_start_idx

Get the index of the first occurrence of a glob symbol.

is_cloud_url

Check if the path is a cloud url.

resolve_homedir

Replaces a “~” in the Path with the home directory.