Trait polars_io::predicates::PhysicalIoExpr

pub trait PhysicalIoExpr: Send + Sync {
    // Required method
    fn evaluate_io(&self, df: &DataFrame) -> PolarsResult<Series>;

    // Provided method
    fn as_stats_evaluator(&self) -> Option<&dyn StatsEvaluator> { ... }
}

Required Methods

fn evaluate_io(&self, df: &DataFrame) -> PolarsResult<Series>

Take a DataFrame and produces a boolean Series that serves as a predicate mask

Provided Methods

fn as_stats_evaluator(&self) -> Option<&dyn StatsEvaluator>

Can take &dyn Statistics and determine of a file should be read -> true or not -> false