polars_io::predicates
Trait PhysicalIoExpr

pub trait PhysicalIoExpr: Send + Sync {
    // Required methods
    fn evaluate_io(&self, df: &DataFrame) -> PolarsResult<Series>;
    fn live_variables(&self) -> Option<Vec<PlSmallStr>>;

    // Provided method
    fn as_stats_evaluator(&self) -> Option<&dyn StatsEvaluator> { ... }
}

Required Methods

fn evaluate_io(&self, df: &DataFrame) -> PolarsResult<Series>

Take a DataFrame and produces a boolean Series that serves as a predicate mask

fn live_variables(&self) -> Option<Vec<PlSmallStr>>

Get the variables that are used in the expression i.e. live variables. This can contain duplicates.

Provided Methods

fn as_stats_evaluator(&self) -> Option<&dyn StatsEvaluator>

Can take &dyn Statistics and determine of a file should be read -> true or not -> false